Fixed warning in VC7
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@990 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/libs/time/sdl/sdltime.c
@@ -22,6 +22,5 @@
 
 Uint32
 SDLWrapper_GetTimeCounter (void) {
-	return SDL_GetTicks () * ONE_SECOND / 1000.0;
+	return (Uint32)(SDL_GetTicks () * ONE_SECOND / 1000.0);
 }
-
